fix: cast wasapi blockalign to usize
- wavefmt.get_blockalign() returns u32 in wasapi 0.23
- Cast to usize at retrieval point so all uses type-check
- Prevents same error recurring at other use sites

+++ loadform/src-tauri/src/audio_capture.rs
@@ -355,7 +355,7 @@ async fn capture_system_audio_windows(
         .get_audiocaptureclient()
         .map_err(|e| format!("Failed to get capture client: {:?}", e))?;
 
-    let blockalign = wavefmt.get_blockalign();
+    let blockalign = wavefmt.get_blockalign() as usize;
 
     audio_client
         .start_stream()
